# docs/developer/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Fyredocs API
  description: |
    Fyredocs is a document processing platform that provides PDF conversion,
    optimization, and organization tools. All responses follow a unified
    envelope format (`APIResponse`) with consistent error handling.

    ## Authentication
    The API supports three authentication methods:
    - **BearerAuth** - JWT access token in the `Authorization: Bearer <token>` header (API clients).
    - **CookieAuth** - `access_token` HttpOnly cookie set on signup/login. Pair with `refresh_token` (Path=`/auth`) for silent rotation via `POST /auth/refresh`.
    - **GuestToken** - `guest_token` HttpOnly cookie issued automatically by the API gateway on first contact when no auth is present. The same cookie is sent on every subsequent request.

    Authenticated users get persistent job storage. Guest users get temporary
    jobs that expire after a configurable TTL.

    ## Presigned Multipart Uploads
    File bytes are uploaded directly to object storage (MinIO/S3) through
    presigned URLs — they never pass through the application services:
    1. `POST /api/upload/init` creates an S3 multipart upload and returns a
       presigned PUT URL for every part.
    2. The browser `PUT`s each part to its presigned URL and collects the
       returned `ETag` response headers.
    3. `GET /api/upload/{uploadId}/parts` re-presigns part URLs to resume an
       interrupted upload or refresh expired URLs.
...
       caller's plan limit.
...

...

...
servers:
...
paths:
...
  /api/orgs:
    get:
      tags: [Organizations]
      summary: List organizations the caller belongs to
      operationId: listOrganizations
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses: { "200": { description: "Orgs, each with the caller's role" } }
    post:
      tags: [Organizations]
      summary: Create an organization (caller becomes owner)
      operationId: createOrganization
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [name], properties: { name: { type: string } } } } }
      responses:
        "201": { description: Organization created }
        "400": { description: Invalid name }
  /api/orgs/{id}:
    get:
      tags: [Organizations]
      summary: Get an org the caller belongs to (with role)
      description: Also used service-to-service by document-service to resolve membership/role for org-scoped requests.
      operationId: getOrganization
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: Organization + caller role }
        "404": { description: "Not found (also returned to non-members)" }
  /api/orgs/{id}/members:
    parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
    get:
      tags: [Organizations]
      summary: List members
      operationId: listMembers
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses: { "200": { description: Members } }
    post:
      tags: [Organizations]
      summary: Add or update a member (admin+)
      operationId: addMember
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [userId, role], properties: { userId: { type: string, format: uuid }, role: { type: string, enum: [admin, editor, viewer] } } } } }
      responses:
        "201": { description: Member added }
        "200": { description: Member updated (already a member) }
        "403": { description: "Caller not admin+, or targeting the owner" }
  /api/orgs/{id}/members/{userId}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: userId, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Organizations]
      summary: Change a member's role (admin+, not the owner)
      operationId: updateMemberRole
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [role], properties: { role: { type: string, enum: [admin, editor, viewer] } } } } }
      responses:
        "200": { description: Role updated }
        "403": { description: "Caller not admin+, or targeting the owner" }
        "404": { description: Member not found }
    delete:
      tags: [Organizations]
      summary: Remove a member (admin+, never the owner)
      operationId: removeMember
      security: [{ BearerAuth: [] }, { CookieAuth: [] }]
      responses:
        "200": { description: Member removed }
        "403": { description: "Caller not admin+, or the owner" }
        "404": { description: Member not found }

  # ==========================================================================
  # Notifications (notification-service :8091)
  # ==========================================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token passed in the `Authorization: Bearer <token>` header.
        Tokens are issued on signup/login and have a configurable TTL (default 8h).

    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored in an HTTP-only, Secure, SameSite=Lax cookie.
        Set automatically on successful signup or login.

    GuestToken:
      type: apiKey
      in: header
      name: X-Guest-Token
      description: |
        Temporary guest identifier for unauthenticated users. Can also be
        sent via the `guest_token` cookie. Allows guests to track their
        processing jobs without creating an account.

  # ---------------------------------------------------------------------------
  # Reusable Response Headers
  # ---------------------------------------------------------------------------
...